requirements.txt
boto3
boto3-stubs[cognito-idp]
pydantic[dotenv]
requests
authlib
cachetools
fastapi[all]
aws-cdk-lib==2.27.0
constructs>=10.0.0,<11.0.0
